Unrolled loops, triple-sort instead of find, separated operator from BlendSeg data structure, found many-loop "bug"
It is becoming increasingly unwieldly to do anything with object_intersection. Will most likely create a new project for QuadEdge meshes that work with Blender, refactor aabb_trees to work with those QE meshes instead, and this should give the power I need. On the other hand, maybe I should just shut up and use it.